treestatus: move non-view helpers out of the API module (bug 1984161)
The Treestatus data-access and update helpers were living alongside the
Ninja API handlers, so unrelated callers such as the landing worker had
to import `is_open` from a views module. Move them to
`lando.treestatus.utils`.
