+++ src/lando/api/legacy/workers/landing_worker.py
@@ -39,7 +39,7 @@ from lando.main.scm import (
     TreeClosed,
 )
 from lando.pushlog.pushlog import PushLog, PushLogForRepo
-from lando.treestatus.views.api import is_open
+from lando.treestatus.utils import is_open
 from lando.utils.config import read_lando_config
 from lando.utils.github import GitHubAPIClient
 from lando.utils.landing_checks import LandingChecks
